// src/types/Fund.ts
export interface Fund {
    last_upd_date: string;
    proj_id: string;
    regis_id: string;
    regis_date?: string;
    cancel_date?: string;
    proj_name_th: string;
    proj_name_en: string;
    proj_abbr_name: string;
    fund_status: string;
    unique_id: string;
    permit_us_investment: string;
    invest_country_flage: string;
}